# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Señaventura",
    "version": "1.0.0",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
...
    "/miembro/": {
      "get": {
        "tags": [
          "Miembro"
        ],
        "description": "Endpoint para obtener todos los miembros.",
        "responses": {
          "200": {
            "description": "Miembros encontrados.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id_miembro": {
                        "type": "integer"
                      },
                      "usuario": {
                        "type": "string"
                      },
                      "es_administrador": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener los miembros.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/miembro/login": {
      "post": {
        "tags": [
          "Miembro"
        ],
        "description": "Endpoint para iniciar sesión como administrador.",
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "autenticado": {
                      "type": "boolean"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Inicio de sesión fallido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "autenticado": {
                      "type": "boolean",
                      "example": false
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al iniciar sesión.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "autenticado": {
                      "type": "boolean",
                      "example": false
                    },
                    "token": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuario": {
                    "type": "string"
                  },
                  "contrasegna": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/miembro/registro": {
      "post": {
        "tags": [
          "Miembro"
        ],
        "description": "Endpoint para registrar un nuevo miembro.",
        "responses": {
          "200": {
            "description": "Registro exitoso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id_miembro": {
                      "type": "integer"
                    },
                    "usuario": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Registro fallido. Usuario ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al registrar.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuario": {
                    "type": "string"
                  },
                  "contrasegna": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/miembro/authToken": {
      "post": {
        "tags": [
          "Miembro"
        ],
        "description": "Endpoint para autenticar un token.",
        "responses": {
          "200": {
            "description": "Token autenticado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "autenticado": {
                      "type": "boolean"
                    },
                    "usuario": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Token no autenticado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "autenticado": {
                      "type": "boolean",
                      "example": false
                    },
                    "usuario": {
                      "type": "string",
                      "example": ""
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al autenticar el token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/miembro/{id}": {
      "put": {
        "tags": [
          "Miembro"
        ],
        "description": "Endpoint para modificar un miembro.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del miembro."
          }
        ],
        "responses": {
          "200": {
            "description": "Miembro modificado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id_miembro": {
                      "type": "integer"
                    },
                    "usuario": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Miembro no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al modificar el miembro.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuario": {
                    "type": "string"
                  },
                  "contrasegna": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Miembro"
        ],
        "description": "Endpoint para eliminar un miembro.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del miembro."
          }
        ],
        "responses": {
          "200": {
            "description": "Miembro eliminado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id_miembro": {
                      "type": "integer"
                    },
                    "usuario": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Miembro no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al eliminar el miembro.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}